+++ src/common/js/startup.js
@@ -3,7 +3,7 @@
  */
 
 Meteor.startup(() => {
-    if( pwixI18n._conf.verbosity & pwixI18n.C.Verbose.DUMP ){
+    if( pwixI18n.configure().verbosity & pwixI18n.C.Verbose.DUMP ){
         console.debug( 'pwixI18n', pwixI18n );
     }
 });